[internal][DateTimePicker] Move useInterceptProps into module
